update pyproject.toml info, add license, rename folder cmr_association_diff to podaac
